+++ react-client/src/routes/my-pets/petId/edit/PetEditIndex.tsx
@@ -1,23 +1,15 @@
 import { SuspenseAwait } from "~/components/await/SuspenseAwait";
-import { Card, DisplayForm, Title } from "~/components/design-system";
 import { Header } from "~/components/design-system/header/Header";
 import { PetModel } from "~/domain/models/pet/PetModel";
 import { PET_PROFILE_FULL_ROUTE } from "~/routes/AppRoutePaths";
 import { invariant } from "~/util/invariant";
-import { PetImageInput } from "../components/PetImageInput";
 import { DiscardConfirmationModal } from "./components/DiscardConfirmationModal";
+import { FormSection } from "./components/FormSection";
 import { usePetEditViewModel } from "./usePetEditViewModel";
 
 export const PetEditIndex = () => {
-  const {
-    form,
-    handleClose,
-    handleReset,
-    onRemoveImage,
-    onSelectImage,
-    onDiscard,
-    petInfoQuery,
-  } = usePetEditViewModel();
+  const { form, handleClose, onDiscard, petInfoQuery, ...formUtility } =
+    usePetEditViewModel();
 
   return <SuspenseAwait resolve={petInfoQuery}>{renderMain}</SuspenseAwait>;
 
@@ -32,10 +24,7 @@ export const PetEditIndex = () => {
           backButtonTo={PET_PROFILE_FULL_ROUTE(pet.id)}
           pageTitle="Edit Pet Profile"
         />
-        <Card padding="xlarge">
-          {renderPetImage(pet)}
-          {renderPetForm()}
-        </Card>
+        <FormSection pet={pet} form={rest} {...formUtility} />
         {isDiscarding && (
           <DiscardConfirmationModal
             onClose={handleClose}
@@ -44,31 +33,5 @@ export const PetEditIndex = () => {
         )}
       </>
     );
-
-    function renderPetImage(pet: PetModel) {
-      return (
-        <>
-          <Title level="h3">Profile Picture</Title>
-          <div className="h-base" />
-          <div className="mb-xxxlarge">
-            <PetImageInput
-              pet={pet}
-              onRemove={onRemoveImage}
-              onSelectFile={onSelectImage}
-            />
-          </div>
-        </>
-      );
-    }
-
-    function renderPetForm() {
-      return (
-        <>
-          <Title level="h3">Pet info</Title>
-          <div className="h-xxlarge" />
-          <DisplayForm {...rest} onReset={handleReset} />
-        </>
-      );
-    }
   }
 };
